use got for fetching manifests
