+++ fetch-mods/test/test.ts
@@ -1,0 +1,4 @@
+import { fetchMods } from "../fetch-mods";
+import modsJson from "./mods.json";
+
+fetchMods(JSON.stringify(modsJson), "ghp_XXX");
